improvments

+++ ParallelMergeSort/src/prog/SortAlgo.java
@@ -6,5 +6,5 @@ import java.util.concurrent.RecursiveTask;
  * Created by o_0 on 2017-03-14.
  */
 public interface SortAlgo <T>{
-    RecursiveTask<double[]> make(double[] array, int threshold);
+    RecursiveTask<float[]> make(float[] array, int threshold);
 }
